Split modbus and cgi, introduce linting and pre-commits
