+++ requirements.txt
@@ -1,0 +1,4 @@
+pymodbus==3.6.6
+requests==2.31.0
+ruff==0.3.3
+pre-commit==3.6.2
