+++ setup.py
@@ -4,7 +4,9 @@ setuptools.setup(
     name="vartastorage",
     version="0.2.7",
     description="VARTA Battery",
-    long_description='With this Python module you can read modbus registers and xml/cgi api values from various VARTA batteries',
+    long_description="""
+        With this Python module you can read modbus registers
+        and xml/cgi api values from various VARTA batteries""",
     url="http://github.com/vip0r/vartastorage",
     author="edAndy",
     author_email="an-dy@gmx.de",
@@ -12,7 +14,7 @@ setuptools.setup(
     package_dir={"": "src"},
     packages=setuptools.find_packages(where="src"),
     zip_safe=False,
-    python_requires='>=3.5',
+    python_requires=">=3.5",
     install_requires=[
         "pymodbus",
         "requests",
